Rename get_attractions.py to attractions_handler.py
