+++ attractions_handler.py
